Refactor CPAL runtime and tuner integration
Introduce an AudioRuntime abstraction and fold tuner functionality into
the CPAL-backed runtime. Replace ringbuf usage with fundsp ThingBuf for
input/spectrum buffers, add a spectrum buffer path, and expose tuner
APIs (start_tuner_only, poll_tuner_spectrum, start/stop_tap_tuner_audio,
update_tuner_config, get_sample_rate).

Move iOS audio session logic into a dedicated audio_session module and
call ensure_configured() at startup. Simplify FFTAnalyzer to push
spectra into a ThingBuf and wire sensor parameter handles into the input
system. Remove legacy tuner runtime and random stream module.

Also update common types and config generation (NodeKeyRegistry helper,
tuner config generation), adjust frontend UI/text (icons, menu, copy),
and drop ringbuf-related dependencies from several Cargo tomls.

+++ crates/audio-system/src/rt/cpal/stream/output.rs
@@ -180,26 +180,26 @@ fn run_output(
 /// If `channels` > 2, the pattern (L,R, 0,0,...) is repeated (simple spread).
 fn write_data<T>(output: &mut [T], channels: usize, next_sample: &mut GenType)
 where
-    T: cpal::SizedSample + cpal::FromSample<f64>,
+    T: cpal::SizedSample + cpal::FromSample<f32>,
 {
     if channels == 0 {
         return;
     }
     for frame in output.chunks_mut(channels) {
         let (l, r) = next_sample();
-        let left: T = T::from_sample(l as f64);
+        let left: T = T::from_sample(l);
 
         if frame.len() == 1 {
             frame[0] = left;
             continue;
         }
 
-        let right: T = T::from_sample(r as f64);
+        let right: T = T::from_sample(r);
         frame[0] = left;
         frame[1] = right;
 
         // Fill any additional channels with silence (or could duplicate).
-        let zero = T::from_sample(0.0f64);
+        let zero = T::from_sample(0.0);
         for slot in frame.iter_mut().skip(2) {
             *slot = zero;
         }
